Added static casts to workaround size_t issue on 31-bit zLinux

+++ src/allocation-tracker.cc
@@ -150,7 +150,8 @@ void AddressToTraceMap::Clear() {
 
 
 void AddressToTraceMap::Print() {
-  PrintF("[AddressToTraceMap (%" V8PRIuPTR "): \n", ranges_.size());
+  PrintF("[AddressToTraceMap (%" V8PRIuPTR "): \n",
+         static_cast<uintptr_t>(ranges_.size()));
   for (RangeMap::iterator it = ranges_.begin(); it != ranges_.end(); ++it) {
     PrintF("[%p - %p] => %u\n", it->second.start, it->first,
         it->second.trace_node_id);
